requirements.txt
# requirements.txt

aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
idna==3.10
iso8601==2.1.0
multidict==6.6.4
mypy_extensions==1.1.0
neispy==5.0.2
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
propcache==0.3.2
pycomcigan==1.4.0
pydantic==2.11.7
pydantic_core==2.33.2
pypika-tortoise==0.6.3
pytz==2025.2
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
starlette==0.47.3
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
